MB-60468: Add OOM-mode to DcpConsumerAckTest
That is the production state where DCP messages are processed but not
acked back to the Producer.
In that state the DcpConsumerTask is scheduled to monitor the OOM state
of the node and acking pending bytes back to the Producer once the node
recovers from OOM.

We miss test coverage for that path.
In this first patch I enable all the existing DcpConsumerAckTest tests
to run in both node healthy/OOM states.
In a subsequent patch I'll expand the coverage for a bunch of DCP
messages that aren't covered by tests at all.

Change-Id: I2b53d692449e88bb3b2ff828bd98f937c04623f2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/204350
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>