MB-59491: Add fast path for bailing out when audit is disabled
Build to build performance comparison identified a 2% drop for
a restore test that was linked to a change to the audit code.
The audit code went from a 2 argument check to a 6 argument check
and the suspicion is that the creation and passing of those extra
arguments increases the cost of operations.

https://review.couchbase.org/c/kv_engine/+/181669

If we can detect early that auditing is disabled, we can avoid the
extra cost.

Change-Id: Ia5267e40268683793f63658f057cbc680cdb137d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/204792
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker