MB-60525: Minimize deadConnections lock scope in ::manageConnections()
The contention on deadConnections between ConnManager and frontend
::disconnect might cause frontend throughput drops in cases where many
DCP connections are being disconnected.

Note: That isn't a golden path. In MB-60525 that happens because
SyncGateway is found opening/closing many DCP connections during the
test. That behaviour isn't expected and is being fixed by SG.

Still, the current code in memcached executes a O(N) for-loop just for
moving every single connection from the deadConnections member to local
toRelease container.

That is suboptimal, and in this patch we achieve the same with a O(1)
swap.

Note: This patch has been tested in a toy-run on MB-60525. This patch
alone doesn't prevent the frontend thoughput drop in that particular
test. Further optimizations in follow-up patches.

Change-Id: I36b40e744d3271940cdef621d368c12d594c27bf
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205005
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>