MB-60497: Add external_auth_slow_duration config parameter
This is the number of seconds to determine a slow operation for an
external authentication request, i.e. the entire duration between
sending and receiving a response.

This parameter allows dynamic configuration of the interval for the
purpose of unit testing. And potentially if we wanted to change the
interval in the future.

Change-Id: I020bd542c84367c17e59cd40e20b053a5f379c4a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205232
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>