MB-60706: Dump connection details at release for DCP
Is unfortunately very common that we get CBSEs for failed rebalances
where the user has triggered the cbcollect after rebalance was aborted
or restarted.

By that we lose the point-in-time DCP view of the problematic state at
rebalance.

This patch adds a connection dump at disconnection. That isn't as
detailed as a full cbcollect, but that dumps a bunch of very useful PiT
information into memcached logs (eg TCP state, userspace libevent
state, etc..).

Eg, before this patch at disconnection we get this in memcached.log:

2024-02-09T14:52:29.336123+01:00 INFO 72: (No Engine) DCP (Producer) eq_dcpq:dcpdrain-9644 - Removing connection [ {"ip":"127.0.0.1","port":61742} - {"ip":"127.0.0.1","port":12000} (<ud>admin</ud>) ]
2024-02-09T14:52:29.375886+01:00 INFO 72: (example) DCP (Producer) eq_dcpq:dcpdrain-9644 - Destroying connection. Created 1.775 s ago. Sent 908 bytes. Noop enabled:true, txInterval:1000ms, pendingRecv:false sendTime:767000000ns s ago, recvTime:767000000ns s ago. Paused 42 times, for 1674 ms total. Details: {"Initializing": {"count":1, "duration":"54 us"},"ReadyListEmpty": {"count":41, "duration":"1674 ms"}}

With this patch we get also this extra information:

2024-02-09T14: 52:29.376035+01:00 INFO 72: Releasing DCP connection: {"sendqueue":{"last":1043000279129875,"size":0,"term":true},"socket_options":{"so_keepalive":8,"so_linger":"off","so_rcvbuf":407672,"so_sndbuf":7456540,"tcp_keepcnt":3,"tcp_keepidle":360,"tcp_keepintvl":10},"total_queued_send":1528,"total_recv":628,"total_send":1528}
Change-Id: Id0e3c510387b3ef80a781ac835091b439ae2b94b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205486
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>