kv_slow_ops_2_gtrace ignore non-Unicode sequences
Our logs can sometimes have invalid bytes sequences in them (for example
due to MB-58598). Specify errors="replace" for opening the input file.

  'replace': Replace with a suitable replacement marker; Python will use
  the official U+FFFD REPLACEMENT CHARACTER for the built-in codecs on
  decoding, and ‘?’ on encoding. Implemented in replace_errors().

We don't care what the original bytes were, so just replace them with a
valid Unicode code point.

Argparse has the same cmdline parsing behaviour as fileinput, but allows
a default to be specified. We have to manually wrap stdin for the
encoding errors parameter to be used. sys.stdin errors is 'strict' on
macOS but 'surrogateecape' on Ubuntu, in my testing. Now, 'replace' is
always used.

Change-Id: I26aa2ab2b3c3e9596b25fbb08d17ecf1427591b1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205379
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>