MB-53681: Add "system" true/false to delete system events
A CreateCollection/ModifyCollection/CreateScope already carries
enough state so that we can determine if they represent a system
collection/scope (one that is not visible without appropriate
privileges).

However the DropCollection/DropScope events cannot be classified
in that way. When a DCP stream without the correct privileges
encounters a DropCollection/DropScope, we must be able to determine
if the event is visible.

This commit updates the FlatBuffer value which those drop events
carries so they are tagged with the system true/false state and
now replicated/stored by all KV nodes. The default for this flag
is false because any drop event created before 7.6.0 will not
be a system event.

Change-Id: Ie05c60d5489b3cbd2b19bbf449580b038581d908
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205573
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>