MB-53681: DCP filtering for connections without SystemCollectionLookup
Full replication is only possible if the connection has the
SystemCollectionLookup privilege, any connection attempting DCP without
must not see events and mutations relating to system collections.

This commit updates the DCP VB::Filter so that a new filter type is
supported, Filter::System.

When constructing the VB::Filter the connection privileges are checked.
If the construction detects that a passthrough (full replication)
filter is being constructed and the connection lacks the privilege a
Filter::System is constructed. This is a filter which initialises the
collection set with all user visible collections. Additionally this
filter is "dynamic" in that new user visible collections extend the
set of collections (similar to a scope filter).

Change-Id: I4dc854b56efc86f80d36d96fe7f794ff88424611
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205575
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>