MB-60806: Remove warning from CouchKVStore::getWithHeader for not found
This logging was added 12 years ago, when we only had value-eviction and
we don't expect negative queries in normal operation. This code is not
on the normal BGFetch path, which batches BGFetches and uses the
getMulti method, so it only gets triggered in rare cases.

In particular, I've noticed it getting triggered from
Manifest::computeDefaultCollectionMaxLegacyDCPSeqno where we check if we
have a ModifyCollection for the _default collection. If we do not have
a ModifyCollection, we log this warning-level message.

Change-Id: Id3ab50a0982d84ebfb051f5c4476b39708b6fad7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206393
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>