MB-60986 Revert "Set Zstd as default for data block compression"
We added Zstd compression as default in Magma as part of MB-52684.
Zstd uses more CPU during compression to provide additional space
saving. This is causing Magma to saturate it's flusher threads
reducing frontend throughput and consuming additional memory.

Reverting to LZ4 which was the old default for Magma to investigate
the issue further.

Change-Id: Ie582a101a873cf6827922e4da8b9c8f19e60ecfe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206574
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>