MB-60879: Expose max HashTable size to prometheus
This will allow customers to set their initial hashtable size to
a large enough size to avoid resizing occurring, at the expense
of higher memory usage.

This is primarily beneficial for customers who want to avoid the
slowness of hashtable resizing, in a workflow where their dataset
grows and shrinks.

Change-Id: Ifb36ccd14cf5b954da5dba61fdf805f9a6438ac5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206782
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>