[BP] MB-60879: Expose max HashTable size to prometheus
This will allow customers to set their initial hashtable size to
a large enough size to avoid resizing occurring, at the expense
of higher memory usage.

This is primarily beneficial for customers who want to avoid the
slowness of hashtable resizing, in a workflow where their dataset
grows and shrinks.

Change-Id: I00e670b2d721074464f607ab59866c3af19f9ce3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207445
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker