MB-61158: Enforce flush_batch_max_bytes
That is the max size in bytes of the flush-batch constructed by the
Flusher and passed to the KVStore for persistence.
Param defaulted to 2GB.

The param enforces the limit at persistence on the single Writer thread run at
EPBucket::flushVBucket(). Ie, multiple Writer threads would still run in
parallel and persist multiple batches, so the actual aggregated IO still
grows linearly with the number of Writers.

For that, the 2GB limitation on the single flush-batch isn't expected to cause
any major performance degradation at persistence.

Change-Id: Ia50016722bdda5f955e176ca19721e7647bf76f7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207185
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>