MB-60861: Add support for structured JSON logs
Added new macros:
 - FATAL_ERROR_CTX
 - LOG_TRACE_CTX
 - LOG_DEBUG_CTX
 - LOG_INFO_CTX
 - LOG_WARNING_CTX
 - LOG_ERROR_CTX
 - LOG_CRITICAL_CTX
 - EP_LOG_TRACE_CTX
 - EP_LOG_DEBUG_CTX
 - EP_LOG_INFO_CTX
 - EP_LOG_WARN_CTX
 - EP_LOG_ERR_CTX
 - EP_LOG_CRITICAL_CTX

The macros accept a message and a cb::logger::JSON object, which gets
appended to the message.

The characters {} are not allowed in the message, and if present, are
replaced with [], to allow simple parsing of the log message (JSON
object always starts at the first open curly brace).

The BucketLogger is also updated, to insert "conn_id" and "bucket" at
the front of the JSON context (for legibility).

Change-Id: Ida83310f0e9704a036b2e996199cc46b2c4db213
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207229
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>