MB-61158: Ensure checkpoint's max size doesn't cross flush_batch_max_bytes
Unlikely to happen in real scenarios as flush_batch_max_bytes is pretty
large (2GB), but we want to be ensure that programmatically.

That is for ensuring that the single checkpoint never build up a
flush-batch that exceeds the max allowed in magma (4GB at the time of
writing).

Change-Id: Ic8b7b261f6328942af147443a17399b394f289af
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207257
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>