MB-54334: Use a per-process unique name for the database
dbnameFromCurrentGTestInfo() tried to be smart on windows and
return a temporary directory for each test case, but unfortunately
that didn't work in the greater picture as the test case start
by _removing_ the test directory to make sure it didn't contain
any "leftovers" from previous runs. By removing the directory
one would open up for another program running on the same
machine to create the _same_ directory and now two processes
would use the same set of files.

Work around the problem by using a fixed name on windows,
but include the process pid in the filename (as that is uniq
as long as the process is running)

Change-Id: Ic25ef53e7e110008bdcdcea1bbb2605b777441ca
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207270
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>