MB-61200: Enforce Checkpoint Quota at DCP replication
The Checkpoint Quota (introduced in 7.1.0) is a hard limit on the
maximum allocation across all vbuckets/checkpoint in a Bucket.
That applies to both frontend and DCP-replication loads.

In MB-31869 (7.6.0) we have erroneously changed the PassiveStream
logic so that DCP-replication bypasses the Checkpoint Quota checks.
We re-introduced those checks here. Covered by unit tests.

Change-Id: Iad234da191c942b145adcddb4b7a973f2657a815
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207338
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker