MB-61174: Don't use 10 iteration counts for unknown users
The patch incorrectly set the iteration counts for unknown users
if the environment variable MEMCACHED_UNIT_TEST wasn't set when the
intention was to reduce the iteration count IF the variable was set.

Also adjust the "default" iteration count for unknown users to 15000
to match the iteration count ns_server is using as of MB-48953.

Change-Id: Ic364a9bef28d2069c3ed9b442a387fe5a92320be
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207362
Well-Formed: Restriction Checker
Reviewed-by: Dan Owen <owend@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>