MB-60431: Replace "unit":"count" with "unit":"none"
Previously, we had a number of metrics that do not have units, but are
showing the unit as count. Because a "unit" field is mandatory, we can
not just delete it. This patch replaces 'count' with 'none'.

Additionally, items that are generated in 'metrics_metadata.json'
which used to have "unit":"count" now no longer a "unit", as
requested by Chris in the MB.

Change-Id: I3f0931fe9e3b6858d1d26d62eabfab41f1b6e63f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207502
Reviewed-by: Dan Owen <owend@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>