MB-61250: Remove completeCompactionExpiryBgFetch deadcode
Early in completeCompactionExpiryBgFetch there are two early returns:

    if (!htRes.committed) {
        return;
    }

    if (!htRes.committed->isTempInitialItem()) {
            return;
    }

Thus the subsequent code can assume htRes.committed is set and the item
is temp-initial.

A subsequent if block is now removed:
   if (htRes.committed && !htRes.committed->isTempItem()

A subsequent code block is removed as it is wrapped with:
   if (!htRes.committed)

Removing this code makes the MB-61250 fix easier to reason about.

Change-Id: Idd47873113d9283f62d0d7e7555200b73d626a07
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207551
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>