MB-61158: Allow >4GB values for flush_batch_max_bytes in EP config
* Quick history recap *

When we started the work for limiting the size of the flush-batch,
magma wasn't expected to become resilient to >4GB-batch in 7.6.0-MP1 /
7.6.1.

In the end magma managed to push their change (MB-61154) into the MP,
so in theory the big batches won't break magma anymore.

We still keep the flush_batch_max_bytes in KV though, as we have
already observed that avoiding huge batches prevents a bunch of magma
subsequent issues (performance in particular).

* Now to this patch *

With MB-61158 we cap flush_batch_max_bytes to 2GB.
Now QE needs to verify MB-61154 for magma, but the only way to verify
that is allowing batches >4GB, which is prevented by KV.

Solution: In this patch we keep the default value for
flush_batch_max_bytes to 2GB, but we remove any param upper bound. So
tests can set the param to any big value for testing purpose.

Change-Id: I104c927a8d26985124f26ef8cd2403ae06439e4a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207568
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>