MB-61220: Format correct vbid in VBCBAdaptor description
Previous code would set currentvb to the next vBucket, then
conditionally pause. If it did pause and the runtime was "slow", there
will be a "Slow runtime" message and the task name will be computed from
getDescription(), which formats the currentvb (which is actually the
next vBucket).

By setting the currentvb member after we've decided to pause, we ensure
that the task's description will reflect the vbid that was just run.

Before the fix, the added unit test sees "no vbucket", "vb:1", "vb:1".
With the fix, the sequence is "no vbucket", "vb:0", "vb:1" as expected.

Change-Id: I16f1d4a22dc02ccf5672cddbc015a76a4335389d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207635
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>