[BP] MB-60431: Replace "unit":"count" with "unit":"none"
Previously, we had a number of metrics that do not have units, but are
showing the unit as count. Because a "unit" field is mandatory, we can
not just delete it. This patch replaces 'count' with 'none'.

Additionally, items that are generated in 'metrics_metadata.json'
which used to have "unit":"count" now no longer a "unit", as
requested by Chris in the MB.

Change-Id: I7e60a437e7304b02ecaeceff06fb8f6b112d702f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207767
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>