MB-61329: test_access_scanner: Skip waiting if already ran
Testing ep_num_access_scanner_skips before the failing check shows that
on success this value is always zero. In the failure case (which was
not locally reproduced) it appears that the access scanner ran/raced
before reading ep_num_access_scanner_skips - then triggering the access
scanner yields no further skips and the test fails/timesout waiting
for skips to go from 4 to 8.

This fix checks the value of ep_num_access_scanner_skips and if not
zero skips the trigger/wait path.

Change-Id: I8c45ec2f2630565b9422bd36641bdf35da86d26f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207821
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>