MB-61159: Log slow/stuck DCP consumers
Introduce two new time intervals causing logging of consumers
using the buffer window:

    log_buffer_full_interval - 5 minutes
    log_buffer_aggreated_full_duration - 30 minutes

Whenever we try to call step() on a producer stream where
the buffer log is full we'll first check if the aggregated
time waiting for the consumer to free up space in the buffer
window exceeds the 30 minute wait time. If that is true
the server will log (every minute) an INFO log message like:

    Total wait time so far for the consumer to free up space
    in the buffer window: <time from cb::time2text>

It'll then check if the _current_ time window its been waiting
for the consumer to free up space exceeds the 5 minutes wait
time. If that is true it'll log (every 5 minutes until the
client makes space available) a WARNING log message like:

    Waited <time from cb::time2text> for the consumer to
    free up space in the buffer window

The first log message will only appear as long as
CB_DEVELOPMENT_ASSERTS is defined (as a long running connection
could eventually get there even if each wait time was <1s)

In order to try to get information for MB-61040 any connection
using a name containing "secidx" reduce those limits to 15 seconds
(continuous blocking) and 1 minute (aggregated value). That would
help add debugging information when the connections seems stuck
(and don't complete in the time window we have logs for)

Change-Id: Ic1f79a21388346d006d00e2eeeecd5a01ce0d5a2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/207154
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>