[BP] MB-60746: Remove obsolete logic in tryToProgressDcpStream
Previously, dcpMaxQSize was dynamically set based on
dcpFlowControlBufferSize. This was introduced in MB-38007, together with
MB-57078, where the size is reduced to 1 MiB if the flow control
buffer was set. The default value from seetings is 1024*1024, and
ns_server does not support a way to set it to something else. Therefore,
removing the conditional check and instead always using the value from
Settings::Instance().getMaxSendQueueSize().

Change-Id: I50e8c59c1d67d3985fd6f1da9dffb1b0d59c4d65
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208629
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker