MB-60746: Remove obsolete logic in tryToProgressDcpStream
Previously, dcpMaxQSize was dynamically set based on
dcpFlowControlBufferSize. This was introduced in MB-38007, together with
MB-57078, where the size is reduced to 1 MiB if the flow control
buffer was set. The default value from seetings is 1024*1024, and
ns_server does not support a way to set it to something else. Therefore,
removing the conditional check and instead always using the value from
Settings::Instance().getMaxSendQueueSize().

Change-Id: Ib2357e86b5e8be9039f5e907e0887778da346dc2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208212
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>