MB-9418: Enable secondary warmup for "on-prem"
Set the configuration so that primary warmup loads no
keys/values (but does load metadata) and so that secondary
warmup loads "100%" of the values (full eviction can be
less than 100%).

As much of the existing tests are running with this configuration
with a few tests running the previous, primary only mode.

Change-Id: Ifffb4ed27f7026a358c3a3aafc3a23fbe3912d4e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208254
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>