MB-61379: FATAL_ERROR should use std::_Exit
It is not safe to use exit() in a multi-threaded program as
we don't know the state of the other threads in the program
and exit may race with these threads and possibly fail
to terminate all of them causing the program to hang around
for a longer time than expected (as observed in the MB).

Use std::_Exit() instead which don't try to run any atexit
handlers or try to nicely shut down the program.

Change-Id: I13398c15fe904dc3d4000dc6db00e84f9cd61cde
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208147
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>