MB-61226: Avoid the branch on ordered for getKey()
getKey() is called from purgeTombstones(), which can run concurrently
with HT operations touching that OSV.

getKey() branches on the ordered bit field, which is packed together
with the datatype field. We can have ActiveItemCompressor storing a
Snappy compressed block concurrently with purgeTombstones() reading the
key from getKey() (and reading the ordered bit).

Since these two operations read/write the same memory location, TSAN
reports this as a data race. However, in this particular case, this is
safe to do (see ticket).

To avoid reading the ordered bit field and the TSAN warning, add a
getKey() shadowing method in OSV, which doesn't read the bit.

Change-Id: I03d5e0ca3a21bdc7e4fd38f6d22c350e78a50b92
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208166
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>