MB-61385: Introduce MaxFutureCasThreshold in HLCT
This patch introduces MaxFutureCasThreshold, which will be used in a
future patch to determine if we would like to accept or reject a set/del
with meta command. The purpose of this parameter is to define a maximum
acceptable threshold of drift at which we will accept a new cas value.
This parameter is also dynamically configured so that users can change
it's value to what suits them best.

Change-Id: Ic6615cbe791a54638c906cf05e1108a8114d7d38
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208246
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>