MB-41351: Add flushUid to manifest.fbs FlatBuffer schemas
Update manifest.fbs (which is used for storing a copy of the most
recently processed node view of the manifest) with a flushUid member.
This is defaulted to 0 so that loading any previously stored manifest
results in a flushUid of 0, which is used to mean no flush has been
requested.

Change-Id: I9d23ea731642ebd7ae0a6216f6d374822b2c1fde
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208653
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>