MB-61627: Reduce locking scope in BackfillManager
A low cost improvement aimed at reducing the scope that the
BackfillManager::lock is held for without a full review of the
complete locking needs of the class (e.g. addition of atomics,
refactoring of other members).

The main aim was to drop the lock before calling functions like
ExecutorPool::wake

Note that wake is resilient to unknown IDs so we're safe to drop the
lock, have another thread remove the task and call wake.

Change-Id: I05491c4b063ef0efc90d505b1a78abe3efb71ba7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/209208
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>