MB-61257: Enforce dcp_backfill_byte_drain_ratio
Core patch for MB-61257.

This fix reverts the backfill-buffer/pausing behaviour to before
https://review.couchbase.org/c/kv_engine/+/180551.

Before that patch: We require that the buffer is "sufficiently drained"
before we can unpause a paused BackfillTask.

After that patch: A paused BackfillTask is unpaused as soon as we jump
1 byte below the backfill buffer limit.
The consequence of that is that the BackfillTask is much more likely to
pause again immediately after pushing a single further item into the
buffer.

As observed in MB-61257, that leads to much more Task snooze/wake that
literally kills the backfill performance.

This patch re-introduces the buffer computations that we changed in
https://review.couchbase.org/c/kv_engine/+/180551.
Differently from that though (where we had hard-coded coefficients for
our computations) now we have the dcp_backfill_byte_drain_ratio param.
By default the param is set to the old hard-coded value (ie, 0.25).

Change-Id: I7cfef7b309d26d6f0018dfe33c945784321d639a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/209841
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>