MB-61902: Minor efficiency tweaks in streamRequest
Remove multiple calls to vb->getHighSeqno as this requires a mutex
to read the value. Read this once into a const.

Remove use of KVBucket::getLastPersistedSeqno as that will
obtain the VBucketPtr (shared_ptr) again, we can get this
data from the vbucket we already have obtained.

Change-Id: I8b90af3aa1d975c0b91e6678102422cf29b8c0eb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210023
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>