MB-34155: Track the type of task that is running
Added GlobalTask::getCurrentTaskType() which returns the type of task
that is executing on this thread (read from a thread_local).

The purpose is to allow code to know what type of thread it is running
on. While this could be implemented at the ExecutorPool level, that
would require duplicating the code for CB3 and Folly, and makes it
slightly more difficult to test.

Change-Id: Ifa26ee8a2d56b0c14e6fa3e0e2b813d1ea16e3c5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210393
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>