MB-34155: Add FileOpsTracker for recording pending IO
The FileOpsTracker class uses a thread-local store for recording
pending IO requests. Any reads and writes that need to be tracked
need to call FileOpsTracker::start with the appropriate request type.
Once the IO has returned, FileOpsTracker::complete must be called to
clear the state.

An external observer can call FileOpsTracker::visitThreads to enumerate
the pending IO requests on all known threads. The external observer can
then filter those by thread type. To detect a stuck Reader or Writer
thread, one can filter for these TaskTypes and check the
FileOp::startTime.

Calling of FileOpsTracker::start/complete will be done from within the
KVStore.

The intention is that should be sufficient to detect a stuck read or
write that would affect a front-end thread in a way that is easy to
reason about (e.g. a writer thread has spent more than 120s in a 4 KiB
write).

Change-Id: Ifaf172d5deccd467f4852399c09cb0459b617d84
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210394
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>