[BP] MB-34155: Track number of bytes since last sync per thread
There is no way to know exactly how many dirty pages for a file need to
be written back for a specific file.

While this is not the exact amount that the fdatasync() call with write
to disk, it is still useful as an approximation, although a thread can
write to several files and then sync, which will further skew the
results.

Change-Id: I6dede2328f8d23d2d46988bb06bd05d263b88ec1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/215168
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>