MB-61736: Fix intermittent failure BgfetchSucceedsUntilMutationWatermark
The unit test BgfetchSucceedsUntilMutationWatermarkt was seen to fail
on CV with configuration magma_persistent_full_eviction. In this test,
memory is configured in such a way that when two BgFetches are run, the
first succeeds and the second fails due to increasing memory above the
mutation watermark.

This patch updates the memory configured to allow an additional 0.3MiB
before reaching the mutation watermark. This will accuont for a
greater overhead for metadata when full eviction is configured.

Change-Id: Id2996f04119f23742473b0fde01459d67c95cb7d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210406
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>