MB-62027: Remove LoadStorageKVPairCallback::purge
The effect of this code if ever invoked is to completely eject
all values. This code was added 14 years ago before full-eviction
and maybe before access log, invoking this function is not desirable.
There is no test coverage either for this code.

If Warmup of any value phase runs out of memory, we will stop Warmup
and be <100% resident which is the desired behaviour.

If KeyDump phase runs out of memory we set the OOMFailure and purge
has no affect anyway as there are no values loaded.

Change-Id: Ic833cd65b577cc35523f79cc7625d54c145a7e34
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210445
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>