MB-62000: Change Warmup to acquire VBucket through a weak_ptr
The VBuckets created by Warmup are the only VBuckets Warmup should
load keys and data into. It is possible (certainly for Secondary
Warmup) that the VBuckets owned by KVBucket are replaced.

This commit removes all Warmup VBucket acquisition via KVBucket, instead
Warmup now has a map of weak_ptr which is used to acquire the VBucket
objects. If Warmup cannot create a shared_ptr, then the VBucket was
removed during Warmup and can now be skipped. If Warmup can obtain a
shared_ptr, it will be the same object that Warmup created and is the
correct VBucket to write data into.

Change-Id: Ied16b385a85ac605cc98579fb90c97b56c6b19fb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210446
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>