MB-60861: Add to_json for common types
Added to_json specialisation used by the JSON log macros macros for the
mcbp Status and EP-Engine status enums, as well as for the Vbid struct.

Note that nlohmann::json will automatically serialise enums to their
underlying integer type, which is not desirable for logging, so we
should ensure that any interesting enums are covered by unit tests.

Change-Id: I16f11904e0719cb8aa2121c4d4004f3b78c5b4f9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210577
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>