[BP] MB-62458: Add magma_WastedSpace to MagmaKVStore stats
Wasted space is important for Magma to track while
debugging issues related to disk space usage by
Magma buckets. Wasted space is the unused space left
in SSTables when an SSTable is appended to as the additional
data results in re-writing of index and footer blocks.
It can result in compactions on it's own and contributes
to Magma fragmentation.

Backport of MB-61585 to 7.2.6

Change-Id: I286aaed1d0b2aa82b772b66da73c26bc9ea6a553
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210784
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211923
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker