[BP] MB-62141: Change ActiveStreamCheckpointProcessorTask to yield by time
BP to 7.6.2 (for 7.6.2-MP3).

- Original commit message -

Prior to this change ActiveStreamCheckpointProcessorTask had an
iteration limit, it would run processing as many vbucket's that were
ready until the iteration count was hit.

This approach is no longer suitable now that long runtimes have been
seen in multi-stream setups.

This commit removes the iteration count (which is limited in that it
didn't count the stream-id loop) and adds a duration check.

The task is configured to run for at most 100 milliseconds, processing
at least 1 ActiveStream and that one stream can still exceed the limit.

The 100 milliseconds was chosen using logs from a real environment
where the runtimes histogram showed that 99.9997% of runtimes were
less than 100 milliseconds.

Note this commit removes any functionality behind the configuration
parameter "dcp_producer_snapshot_marker_yield_limit" and replaces
it with "dcp_producer_processor_run_duration_us". A parameter that
defines the max microseconds for the task. Microseconds giving lots
of flexibility if ever the need for sub millisecond limits.

Change-Id: Ib30eb00a56001cb02ee50496c15c92b278124c10
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220818
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>