MB-62152: Fix the dcp_backfill_byte_drain_ratio=1.0 case
Param introduced in MB-61257.

Our logic isn't resilient to that param set to 1.0. By that, the
backfill->memory transition on a stream might not be properly
processed. That might result in DCP cursors stuck at the start of
open checkpoints when backfill completes and the stream goes in-memory.

Change-Id: I417a3e4443e039fc72de1e458936f0fb6ca4872f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210814
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker