MB-62118: vb_ht_max_size should not have bytes unit
vb_ht_max_size is the number of slots in the hashtable.
Previously, the unit was incorrectly set to bytes. This
then adds a suffix to the end of the metric name:
'vb_ht_max_size_bytes' and can be misleading.

This patch removes the unit and suffix from the metric
name.

Change-Id: Id183bd6e394076b5cb0c95c92c1b6ffa64d5c2f8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/210859
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>