MB-62223: Replica must forceMaxCas
A replica vbucket must not set the max_cas to the highest observed
cas, that would mean a mismatch with the active in the case where
an active has a non-monotonic CAS w.r.t to document seqnos. This
can occur in the case where a vbucket has been fixed (forceMaxCas)
due to a prior issue.

The correct thing for a replica todo is always set max_cas to equal
the cas of the high-seqno, as that would be the same max_cas as the
active.

There is still the possibility that the active vbucket has had a
forceMaxCas, yet no data transfers, allowing a difference, however all
forceMaxCas guidelines ask to change all vbuckets (Active+Replicas).

Change-Id: I8287394f259b7be4a773fd7f8d4dd33d3870b427
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211065
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>