[BP] MB-60382: Fix intermittent GetAllKeys Failure
We should wait till all upserts have persisted before initiating a
GetAllKeys command. The test was seen to fail when a document was
still pending and has not yet persisted. AllKeysCallback skips
durability-prepared keys, and on machines where the flusher is slow,
those keys are not returned.

This patch retrieves the last seqno and persists all seqnos up
to and including this one, ensuring that no key remains in the
prepared state.

Change-Id: I479480f5c56444658b3d89f3f7423bf00bd96a6d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211798
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>